+++ ui/src/pages/DetailPage.jsx
