+++ ui/src/pages/ListPage.jsx
@@ -3,7 +3,7 @@ import Container from '@material-ui/core/Container';
 import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import FlowList from './FlowList';
+import FlowList from '../components/FlowList';
 
 export default function ListView() {
   return (
